Move filterbuilder to frontend-core
